Update README.md
This is just the top 1/3 of my tribute page. The rest will consist of bullet points with some text information and an external link to a website with further information on Kobe Bryant.

I am also hoping to separate the bottom 2/3 of the page into 2 uneven columns to have the bullet points on the left side (in the larger column) and an embedded youtube video showing a Kobe mixtape on the right side (smaller column).